refactored the code a little bit for isolating useState
