+++ src/components/Projects/FeaturedProjects.tsx
@@ -1,12 +1,7 @@
-'use client';
-import { useState } from 'react';
-import ProjectCardCompact from './Cards/ProjectCardCompact';
+// 'use client'; // No longer needed at the top of this file
 import BarebonesCard from '../cards/BarebonesCard';
-import ProjectModal from './ProjectModal';
 import Link from 'next/link';
-// import Image from 'next/image';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faUser } from '@fortawesome/free-solid-svg-icons';
+import ClientSideProjects from './ClientSideProjects';
 
 const mockProjectDetails = [
   {
@@ -19,27 +14,6 @@ const mockProjectDetails = [
 ];
 
 export default function FeaturedProjects() {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [selectedProject, setSelectedProject] = useState<{
-    id: number;
-    title: string;
-    description: string;
-  } | null>(null);
-
-  const openModal = (project: {
-    id: number;
-    title: string;
-    description: string;
-  }) => {
-    setSelectedProject(project);
-    setIsModalOpen(true);
-  };
-
-  const closeModal = () => {
-    setIsModalOpen(false);
-    setSelectedProject(null);
-  };
-
   return (
     <BarebonesCard>
       <div className="flex flex-col">
@@ -47,16 +21,8 @@ export default function FeaturedProjects() {
           <span className="">Featured Projects</span>
         </h1>
         <div className="gap-2 my-2 flex flex-wrap lg:flex-nowrap">
-          {/* Render Project Cards */}
-          {mockProjectDetails.map((project) => (
-            <div
-              key={project.id}
-              onClick={() => openModal(project)}
-              className="cursor-pointer"
-            >
-              <ProjectCardCompact className="w-full lg:w-auto" />
-            </div>
-          ))}
+          {/* Pass mock project details to the client-side component */}
+          <ClientSideProjects projects={mockProjectDetails} />
         </div>
         <div className="ml-auto text-right">
           <Link
@@ -67,87 +33,6 @@ export default function FeaturedProjects() {
           </Link>
         </div>
       </div>
-
-      {/* Modal for Project Details */}
-      <ProjectModal isOpen={isModalOpen} onClose={closeModal}>
-        {selectedProject && (
-          <div className="overflow-scroll max-h-[calc(100vh-100px)]">
-            <h2 className="font-bold text-xl">{selectedProject.title}</h2>
-            <div className="flex flex-col md:flex-row gap-2">
-              <div className="w-full md:w-1/3 flex flex-col gap-2">
-                <div className="relative flex items-center justify-center w-full h-full rounded-lg bg-red-400">
-                  <FontAwesomeIcon icon={faUser} />
-                </div>
-                <div className="relative flex items-center justify-center w-full h-full rounded-lg bg-red-400">
-                  <FontAwesomeIcon icon={faUser} />
-                </div>
-              </div>
-
-              <div className="w-full md:w-2/3 ">
-                {/* <p>{selectedProject.description}</p> */}
-                <div className="mb-2">
-                  <h3 className="font-bold text-lg">What is It?</h3>
-                  <p>
-                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                    Quis, unde nulla voluptatibus alias iste incidunt nihil
-                    ipsam molestiae sint nisi eveniet, dolorum in officia magnam
-                    doloremque corrupti autem repudiandae provident. Lorem ipsum
-                    dolor sit amet consectetur adipisicing elit. Quis, unde
-                    nulla voluptatibus alias iste incidunt nihil ipsam molestiae
-                    sint nisi eveniet, dolorum in officia magnam doloremque
-                    corrupti autem repudiandae provident. ipsam molestiae sint
-                    nisi eveniet, dolorum in officia magnam doloremque corrupti
-                    autem repudiandae provident. Lorem ipsum dolor sit amet
-                    consectetur adipisicing elit. Quis, unde nulla voluptatibus
-                    alias iste incidunt nihil ipsam molestiae sint nisi eveniet,
-                    dolorum in officia magnam doloremque corrupti autem
-                    repudiandae provident.ipsam molestiae sint nisi eveniet,
-                    dolorum in officia magnam doloremque corrupti autem
-                    repudiandae provident. Lorem ipsum dolor sit amet
-                    consectetur adipisicing elit. Quis, unde nulla voluptatibus
-                    alias iste incidunt nihil ipsam molestiae sint nisi eveniet,
-                    dolorum in officia magnam doloremque corrupti autem
-                    repudiandae provident.ipsam molestiae sint nisi eveniet,
-                    dolorum in officia magnam doloremque corrupti autem
-                    repudiandae provident. Lorem ipsum dolor sit amet
-                    consectetur adipisicing elit. Quis, unde nulla voluptatibus
-                    alias iste incidunt nihil ipsam molestiae sint nisi eveniet,
-                    dolorum in officia magnam doloremque corrupti autem
-                    repudiandae provident.
-                  </p>
-                </div>
-                <div className="mb-2">
-                  <h3 className="font-bold text-lg">Why I did It?</h3>
-                  <p>
-                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                    Quis, unde nulla voluptatibus alias iste incidunt nihil
-                    ipsam molestiae sint nisi eveniet, dolorum in officia magnam
-                    doloremque corrupti autem repudiandae provident.
-                  </p>
-                </div>
-                <div className="mb-2">
-                  <h3 className="font-bold text-lg">What I Learned?</h3>
-                  <p>
-                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                    Quis, unde nulla voluptatibus alias iste incidunt nihil
-                    ipsam molestiae sint nisi eveniet, dolorum in officia magnam
-                    doloremque corrupti autem repudiandae provident.
-                  </p>
-                </div>
-                <div className="mb-2">
-                  <h3 className="font-bold text-lg">Future</h3>
-                  <p>
-                    Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                    Quis, unde nulla voluptatibus alias iste incidunt nihil
-                    ipsam molestiae sint nisi eveniet, dolorum in officia magnam
-                    doloremque corrupti autem repudiandae provident.
-                  </p>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-      </ProjectModal>
     </BarebonesCard>
   );
 }
